+++ actionControllers/ProjectController.ts
@@ -2,6 +2,7 @@ import { ActionContext, Request, Response } from '@frontastic/extension-types';
 import { getCurrency, getLocale } from '../utils/Request';
 import { ProjectApi } from '../apis/ProjectApi';
 import handleError from '@Commerce-commercetools/utils/handleError';
+import extractRegionFromCommercetoolsHostUrl from '@Commerce-commercetools/utils/extractRegionFromCommercetoolsHostUrl';
 
 type ActionHook = (request: Request, actionContext: ActionContext) => Promise<Response>;
 
@@ -15,9 +16,11 @@ export const getProjectSettings: ActionHook = async (request: Request, actionCon
 
     const project = await projectApi.getProjectSettings();
 
+    const region = extractRegionFromCommercetoolsHostUrl(actionContext.frontasticContext);
+
     const response: Response = {
       statusCode: 200,
-      body: JSON.stringify(project),
+      body: JSON.stringify({ ...project, region }),
       sessionData: {
         ...projectApi.getSessionData(),
       },