+++ utils/CartFetcher.ts
@@ -1,10 +1,11 @@
 import { Request } from '@frontastic/extension-types';
 import { Cart } from '@Types/cart/Cart';
 import { CartApi } from '../apis/CartApi';
+import { ExternalError } from '@Commerce-commercetools/errors/ExternalError';
 
 export class CartFetcher {
   static async fetchCart(cartApi: CartApi, request: Request): Promise<Cart> {
-    const cart = await this.fetchCartFromSession(cartApi, request);
+    const cart = await this.fetchActiveCartFromSession(cartApi, request);
 
     if (cart) {
       return cart;
@@ -13,12 +14,19 @@ export class CartFetcher {
     return await cartApi.getAnonymous();
   }
 
-  static async fetchCartFromSession(cartApi: CartApi, request: Request): Promise<Cart | undefined> {
+  static async fetchActiveCartFromSession(cartApi: CartApi, request: Request): Promise<Cart | undefined> {
     if (request.sessionData?.cartId !== undefined) {
       try {
-        return await cartApi.getActiveCartById(request.sessionData.cartId);
+        const cart = await cartApi.getById(request.sessionData.cartId);
+        if (cartApi.assertCartIsActive(cart)) {
+          return cart;
+        }
       } catch (error) {
-        console.info(`It was not possible to fetch the cart ${request.sessionData.cartId}. ${error}`);
+        // A ExternalError might be thrown if the cart does not exist or belongs to a different business unit,
+        // in which case we should create a new cart.
+        if (!(error instanceof ExternalError)) {
+          throw error;
+        }
       }
     }